api/requirements.txt
fastapi==0.110.3
uvicorn[standard]==0.29.0
motor==3.4.0
pymongo==4.8.0
mysql-connector-python==8.3.0
